1. swagger 수정 2. 프로젝트 구조 개선
